refactor(ui): move selection and focus out of model.go
Cursor movement, panel focus, and selection share one question: what does the
next key act on. Tag expansion belongs with them because selecting a tag selects
its services.

Declarations move unchanged.
